requirements.txt
anyio==4.6.0
certifi==2024.8.30
colorama==0.4.6
exceptiongroup==1.2.2
Faker==33.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
load-dotenv==0.1.0
loguru==0.7.2
pillow==11.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.6
six==1.16.0
sniffio==1.3.1
typing_extensions==4.12.2
validators==0.34.0
win32-setctime==1.1.0
